tools/perf/tests: fix test_arm_spe.sh signal case issues
This patch aims to address the shellcheck issue in test_arm_spe.sh
for the mixed-case warnings in the signal names for the trap command.
This patch fixes the issue by capitalizing the "EXIT", "TERM" and
"INIT" signals given as inputs to the trap required, enhancing
code-readability and following linux coding best-practices.

Signed-off-by: Abhirup Deb <abhirup.deb@linux.vnet.ibm.com>i
Signed-off-by: Ojaswin Mujoo <ojaswin.mujoo@ibm.com>
Signed-off-by: Piyush Sachdeva <Piyush.Sachdeva@ibm.com>
Signed-off-by: Mukesh Chaurasiya <mukesh.chaurasiya@ibm.com>